+++ packages/angular/tests/index.test.ts
@@ -1,6 +1,7 @@
 /**
  * @jest-environment jsdom
  */
+import '@angular/compiler'
 
 import { NavigationEnd, NavigationStart } from '@angular/router'
 import type { Router } from '@angular/router'