+++ docs/RELEASES.md
@@ -30,6 +30,7 @@ You are now ready to make the release. Releases are done using Docker and Docker
 
 - You will need to clone the repository and have Docker running on your local machine
 - Ensure you are logged in to npm and that you have access to publish any packages in the `@bugsnag` namespace
+- Generate a [granular access token](https://www.npmjs.com/settings/{username}/tokens/granular-access-tokens/new) on NPM as a fallback mechanism when 2FA is not feasible and store it somewhere secure
 - If there are changes to the browser client, ensure you have permissions to upload to the CDN bucket
 - Ensure your `.gitconfig` file in your home directory is configured to contain your name and email address
 - Generate a [personal access token](https://github.com/settings/tokens/new) on GitHub and store it somewhere secure
@@ -49,6 +50,7 @@ Ensure the following environment variables are set:
 - AWS_ACCESS_KEY_ID
 - AWS_SECRET_ACCESS_KEY
 - AWS_SESSION_TOKEN
+- NPM_TOKEN
 - AWS_REGION
 - BUCKET_NAME
 - DISTRIBUTION_ID
@@ -61,7 +63,7 @@ RELEASE_BRANCH=main \
   docker compose run release
 ```
 
-This process is interactive and will require you to confirm that you want to publish the changed packages. It will also prompt for 2FA.
+This process is interactive and will require you to confirm that you want to publish the changed packages.
 
 Browser bundles will be automatically uploaded to the CDN if they have changed.